Put PollOption back in a separate file, its docstrings must be visible
