+++ pyrogram/client/types/messages_and_media/__init__.py
@@ -27,6 +27,7 @@ from .messages import Messages
 from .photo import Photo
 from .photo_size import PhotoSize
 from .poll import Poll
+from .poll_option import PollOption
 from .sticker import Sticker
 from .user_profile_photos import UserProfilePhotos
 from .venue import Venue
